Added OperationResponseFuture type
